Добавление валидации на работу с расписанием (добавлять запись может только доктор сам себе, либо админ любому

+++ private-service/src/main/java/ru/bmstu/appointment/privateservice/controller/ScheduleController.java
@@ -11,11 +11,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.bmstu.appointment.commonmodel.dto.ScheduleRequest;
 import ru.bmstu.appointment.commonmodel.dto.ScheduleResponse;
+import ru.bmstu.appointment.commonmodel.model.Doctor;
 import ru.bmstu.appointment.commonmodel.model.Schedule;
+import ru.bmstu.appointment.commonmodel.model.User;
 import ru.bmstu.appointment.commonmodel.repository.DoctorRepository;
 import ru.bmstu.appointment.commonmodel.repository.ScheduleRepository;
+import ru.bmstu.appointment.commonmodel.repository.UserRepository;
 import ru.bmstu.appointment.commonmodel.utils.ScheduleMapping;
+import ru.bmstu.appointment.privateservice.config.jwt.JwtProvider;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -35,7 +40,13 @@ public class ScheduleController {
     private DoctorRepository doctorRepository;
 
     @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
     private ScheduleMapping scheduleMapping;
+
+    @Autowired
+    private JwtProvider jwtProvider;
 
     @SecurityRequirements
     @Operation(summary = "Расписание врачей в зависимости от заданных фильтров")
@@ -63,10 +74,18 @@ public class ScheduleController {
             content = {@Content(mediaType = "application/json",
                     schema = @Schema(implementation = ScheduleResponse.class))})})
     @PostMapping("/add")
-    public ScheduleResponse add(@RequestBody @Valid ScheduleRequest request) throws ParseException {
+    public ScheduleResponse add(@RequestBody @Valid ScheduleRequest scheduleRequest, HttpServletRequest request)
+            throws ParseException {
+        String username = jwtProvider.getLoginFromToken(request.getHeader("Authorization").substring(7));
+        User user = userRepository.findByUsername(username);
         Schedule schedule = new Schedule();
-        schedule.setDoctor(doctorRepository.getOne(request.getDoctor()));
-        return getScheduleResponse(request, schedule);
+        if (user.getRole().getName().equals("ROLE_DOCTOR")) {
+            Doctor doctor = doctorRepository.getDoctorByUser(userRepository.findByUsername(username));
+            schedule.setDoctor(doctor);
+        } else {
+            schedule.setDoctor(doctorRepository.getOne(scheduleRequest.getDoctor()));
+        }
+        return getScheduleResponse(scheduleRequest, schedule);
     }
 
     private ScheduleResponse getScheduleResponse(@RequestBody @Valid ScheduleRequest request, Schedule schedule) throws ParseException {
@@ -82,10 +101,16 @@ public class ScheduleController {
             content = {@Content(mediaType = "application/json",
                     schema = @Schema(implementation = ScheduleResponse.class))})})
     @PutMapping("/update/{id}")
-    public ScheduleResponse update(@RequestBody @Valid ScheduleRequest request, @PathVariable UUID id) throws ParseException {
+    public ScheduleResponse update(@RequestBody @Valid ScheduleRequest scheduleRequest, HttpServletRequest request,
+                                   @PathVariable UUID id) throws ParseException {
         if (scheduleRepository.existsById(id)) {
             Schedule schedule = scheduleRepository.getOne(id);
-            return getScheduleResponse(request, schedule);
+            String username = jwtProvider.getLoginFromToken(request.getHeader("Authorization").substring(7));
+            User user = userRepository.findByUsername(username);
+            if (schedule.getDoctor().getUser().getUsername().equals(username) ||
+                    user.getRole().getName().equals("ROLE_ADMIN")) {
+                return getScheduleResponse(scheduleRequest, schedule);
+            }
         }
         return null;
     }
@@ -93,10 +118,15 @@ public class ScheduleController {
 
     @Operation(summary = "Удаление записи", description = "Доступно для пользователей с ролью DOCTOR")
     @DeleteMapping("/delete/{id}")
-    public String delete(@PathVariable UUID id) {
+    public String delete(@PathVariable UUID id, HttpServletRequest request) {
         if (scheduleRepository.existsById(id)) {
-            scheduleRepository.deleteById(id);
-            return "OK";
+            String username = jwtProvider.getLoginFromToken(request.getHeader("Authorization").substring(7));
+            User user = userRepository.findByUsername(username);
+            if (scheduleRepository.getOne(id).getDoctor().getUser().getUsername().equals(username) ||
+                    user.getRole().getName().equals("ROLE_ADMIN")) {
+                scheduleRepository.deleteById(id);
+                return "OK";
+            }
         }
         return null;
     }
